Update Tailwind Filters
- Update folders
- Update documentation (prepare for IDE, TypeDoc)
- Fix ambiguity problem with IDE and TypeDoc
