+++ src/tailwind-filters/tailwind-filters-blur-types.ts
@@ -1,0 +1,127 @@
+// noinspection JSUnusedGlobalSymbols
+
+/**
+ * @group Filters
+ * @category Blur
+ * @css
+ * ```css
+ * filter: blur(0);
+ * ```
+ * @description You can refer to the official [tailwind documentation](https://tailwindcss.com/docs/blur)
+ * @description You can refer to the [typedoc](https://git-init-wesley.github.io/tailwindcss-types/docs/types/blur_none.html)
+ * @since Tailwind v3.2.7
+ */
+export type blur_none = 'blur-none'
+
+/**
+ * @group Filters
+ * @category Blur
+ * @css
+ * ```css
+ * filter: blur(4px);
+ * ```
+ * @description You can refer to the official [tailwind documentation](https://tailwindcss.com/docs/blur)
+ * @description You can refer to the [typedoc](https://git-init-wesley.github.io/tailwindcss-types/docs/types/blur_sm.html)
+ * @since Tailwind v3.2.7
+ */
+export type blur_sm = 'blur-sm'
+
+/**
+ * @group Filters
+ * @category Blur
+ * @css
+ * ```css
+ * filter: blur(8px);
+ * ```
+ * @description You can refer to the official [tailwind documentation](https://tailwindcss.com/docs/blur)
+ * @description You can refer to the [typedoc](https://git-init-wesley.github.io/tailwindcss-types/docs/types/blur_md.html)
+ * @since Tailwind v3.2.7
+ */
+export type blur = 'blur'
+
+/**
+ * @group Filters
+ * @category Blur
+ * @css
+ * ```css
+ * filter: blur(12px);
+ * ```
+ * @description You can refer to the official [tailwind documentation](https://tailwindcss.com/docs/blur)
+ * @description You can refer to the [typedoc](https://git-init-wesley.github.io/tailwindcss-types/docs/types/blur_md.html)
+ * @since Tailwind v3.2.7
+ */
+export type blur_md = 'blur-md'
+
+/**
+ * @group Filters
+ * @category Blur
+ * @css
+ * ```css
+ * filter: blur(16px);
+ * ```
+ * @description You can refer to the official [tailwind documentation](https://tailwindcss.com/docs/blur)
+ * @description You can refer to the [typedoc](https://git-init-wesley.github.io/tailwindcss-types/docs/types/blur_lg.html)
+ * @since Tailwind v3.2.7
+ */
+export type blur_lg = 'blur-lg'
+
+/**
+ * @group Filters
+ * @category Blur
+ * @css
+ * ```css
+ * filter: blur(24px);
+ * ```
+ * @description You can refer to the official [tailwind documentation](https://tailwindcss.com/docs/blur)
+ * @description You can refer to the [typedoc](https://git-init-wesley.github.io/tailwindcss-types/docs/types/blur_xl.html)
+ * @since Tailwind v3.2.7
+ */
+export type blur_xl = 'blur-xl'
+
+/**
+ * @group Filters
+ * @category Blur
+ * @css
+ * ```css
+ * filter: blur(40px);
+ * ```
+ * @description You can refer to the official [tailwind documentation](https://tailwindcss.com/docs/blur)
+ * @description You can refer to the [typedoc](https://git-init-wesley.github.io/tailwindcss-types/docs/types/blur_2xl.html)
+ * @since Tailwind v3.2.7
+ */
+export type blur_2xl = 'blur-2xl'
+
+/**
+ * @group Filters
+ * @category Blur
+ * @css
+ * ```css
+ * filter: blur(64px);
+ * ```
+ * @description You can refer to the official [tailwind documentation](https://tailwindcss.com/docs/blur)
+ * @description You can refer to the [typedoc](https://git-init-wesley.github.io/tailwindcss-types/docs/types/blur_3xl.html)
+ * @since Tailwind v3.2.7
+ */
+export type blur_3xl = 'blur-3xl'
+
+/**
+ * @group Filters
+ * @category Blur
+ * @example
+ * ```ts
+ * let value: Blur //Only tailwindcss types
+ * let value: string | Blur //Your custom tailwindcss with only tailwindcss type
+ * ```
+ * @description You can refer to the official [tailwind documentation](https://tailwindcss.com/docs/blur)
+ * @description You can refer to the [typedoc](https://git-init-wesley.github.io/tailwindcss-types/docs/types/Blur.html)
+ * @since Tailwind v3.2.7
+ */
+export type Blur =
+    blur_none
+    | blur_sm
+    | blur
+    | blur_md
+    | blur_lg
+    | blur_xl
+    | blur_2xl
+    | blur_3xl
