+++ src/tailwind-filters/tailwind-filters-brightness-types.ts
@@ -1,0 +1,24 @@
+/**
+ * @group Filters
+ * @category Brightness
+ * @example
+ * ```ts
+ * let value: Brightness //Only tailwindcss types
+ * let value: string | Brightness //Your custom tailwindcss with only tailwindcss type
+ * ```
+ * @description You can refer to the official [tailwind documentation](https://tailwindcss.com/docs/brightness)
+ * @description You can refer to the [typedoc](https://git-init-wesley.github.io/tailwindcss-types/docs/types/Brightness.html)
+ * @since Tailwind v3.2.7
+ */
+export type Brightness =
+    'brightness-0'
+    | 'brightness-50'
+    | 'brightness-75'
+    | 'brightness-90'
+    | 'brightness-95'
+    | 'brightness-100'
+    | 'brightness-105'
+    | 'brightness-110'
+    | 'brightness-125'
+    | 'brightness-150'
+    | 'brightness-200'
